Accounting for quasar_ids being strings

+++ package_data_for_lephare.py
@@ -6,7 +6,7 @@ called glq_quasar_data.in
 import numpy as np
 import pandas as pd
 
-quasar_ids = np.load("./data/processed/quasar_ids.npz")["ids"]
+quasar_ids = np.load("./data/processed/quasar_ids.npz")["ids"].astype(int)
 
 df = pd.read_csv("./data/processed/cut_crossmatched_objects.csv", index_col=0).loc[
     quasar_ids
